The likelihood ratio confidence interval for the binomial probability

The likelihood ratio confidence interval for the binomial probability. Described in Chapter 2 "The 1x2 Table and the Binomial Distribution"

LR_CI_1x2(X, n, alpha = 0.05)

Arguments

X

the number of successes

n

the total number of observations

alpha

the nominal level, e.g. 0.05 for 95# CIs

Value

An object of the contingencytables_result class, basically a subclass of base::list(). Use the utils::str() function to see the specific elements returned.

Examples

LR_CI_1x2(singh_2010["1st", "X"], singh_2010["1st", "n"])
#> The likelihood ratio CI: estimate = 0.4690 (95% CI 0.4269 to 0.5115)
LR_CI_1x2(singh_2010["2nd", "X"], singh_2010["2nd", "n"])
#> The likelihood ratio CI: estimate = 0.4951 (95% CI 0.4470 to 0.5433)
LR_CI_1x2(singh_2010["3rd", "X"], singh_2010["3rd", "n"])
#> The likelihood ratio CI: estimate = 0.6168 (95% CI 0.5417 to 0.6883)
with(singh_2010["4th", ], LR_CI_1x2(X, n)) # alternative syntax
#> The likelihood ratio CI: estimate = 0.7333 (95% CI 0.5936 to 0.8472)
LR_CI_1x2(ligarden_2010["X"], ligarden_2010["n"])
#> The likelihood ratio CI: estimate = 0.8125 (95% CI 0.5828 to 0.9497)