Statistical Analysis of Contingency tables

Statistical Analysis of Contingency Tables is an invaluable tool for statistical inference in contingency tables. It covers effect size estimation, confidence intervals, and hypothesis tests for the binomial and the multinomial distributions, unpaired and paired 2x2 tables, rxc tables, ordered rx2 and 2xc tables, paired cxc tables, and stratified tables. This package provides functions that accompany the "Statistical Analysis of Contingency Tables" book by Fagerland et. al. <ISBN 9781466588172>.

References

Fagerland MW, Lydersen S, Laake P (2017) Statistical Analysis of Contingency Tables. Chapman & Hall/CRC, Boca Raton, FL

https://contingencytables.com/

https://www.routledge.com/Statistical-Analysis-of-Contingency-Tables/Fagerland-Lydersen-Laake/p/book/9781466588172

https://ocbe-uio.github.io/contingencytables/

See also

Author

Maintainer: Waldir Leoncio w.l.netto@medisin.uio.no

Authors:

Morten Wang Fagerland morten.fagerland@medisin.uio.no

Other contributors:

Stian Lydersen [contributor]

Petter Laake [contributor]

Ole Christian Lingjærde [translator]

Brad J. Biggerstaff [contributor]