The Paired 2x2 table tests

The Paired 2x2 table tests

the_paired_2x2_table_tests(n, gamma = 1e-04, num_pi_values = 1000L)

Arguments

n

frequency matrix

gamma

parameter for the Berger and Boos procedure

num_pi_values

number of values to use in the partition of the nuisance parameter space (default=1000)

Value

NULL. This function should be called for its printed output.

Examples

the_paired_2x2_table_tests(bentur_2009)
#> 
#> H_0: pi_1+ = pi_+1  vs  H_A: pi_1+ ~= pi_+1
#> 
#> Estimate of pi_1+: 2/21 = 0.095
#> Estimate of pi_+1: 8/21 = 0.381
#> 
#> Test                             P-value  (test statistic)
#> ---------------------------------------------------------
#> McNemar asymptotic               0.0339   (Z = -2.121)
#> McNemar asymptotic w/CC          0.0771   (Z = 1.768)
#> McNemar exact conditional        0.0703
#> McNemar mid-P                    0.0391
#> McNemar exact unconditional      0.0353
#> McNemar exact unconditional*     0.0354
#> ---------------------------------------------------------
#> *gamma = 0.0001    
the_paired_2x2_table_tests(cavo_2012, gamma = 0, num_pi_values = 10)
#> 
#> H_0: pi_1+ = pi_+1  vs  H_A: pi_1+ ~= pi_+1
#> 
#> Estimate of pi_1+: 65/161 = 0.404
#> Estimate of pi_+1: 75/161 = 0.466
#> 
#> Test                             P-value  (test statistic)
#> ---------------------------------------------------------
#> McNemar asymptotic               0.0330   (Z = -2.132)
#> McNemar asymptotic w/CC          0.0550   (Z = 1.919)
#> McNemar exact conditional        0.0525
#> McNemar mid-P                    0.0347
#> McNemar exact unconditional*     0.0339
#> ---------------------------------------------------------
#> *gamma = 0         
the_paired_2x2_table_tests(ezra_2010, gamma = 0, num_pi_values = 20)
#> 
#> H_0: pi_1+ = pi_+1  vs  H_A: pi_1+ ~= pi_+1
#> 
#> Estimate of pi_1+: 32/102 = 0.314
#> Estimate of pi_+1: 9/102 = 0.088
#> 
#> Test                             P-value  (test statistic)
#> ---------------------------------------------------------
#> McNemar asymptotic               0.0000   (Z = 4.426)
#> McNemar asymptotic w/CC          0.0000   (Z = 4.234)
#> McNemar exact conditional        0.0000
#> McNemar mid-P                    0.0000
#> McNemar exact unconditional*     0.0000
#> ---------------------------------------------------------
#> *gamma = 0